Renamed genesetdatatypeagent.ts to genesetdatatype.ts
